# proyecto_Lotuz/src/main/resources/static/openapi.json
{"openapi":"3.0.3","info":{"title":"Lotuz API","version":"1.0.0","description":"API del backend de Lotuz para autenticación, catálogo, perfil, checkout y administración."},"servers":[{"url":"http://localhost:8080","description":"Servidor local de desarrollo"}],"paths":{"/api/auth/register":{"post":{"tags":["Autenticación"],"summary":"Registrar nuevo usuario cliente","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Usuario"}}}},"responses":{"201":{"description":"Usuario creado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Usuario"}}}},"400":{"description":"Datos inválidos"}}}},"/api/auth/login":{"post":{"tags":["Autenticación"],"summary":"Iniciar sesión","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}}},"responses":{"200":{"description":"Inicio de sesión exitoso","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Usuario"}}}},"401":{"description":"No autorizado"}}}},"/api/productos":{"get":{"tags":["Productos"],"summary":"Listar productos públicos","parameters":[{"name":"terminoBusqueda","in":"query","schema":{"type":"string"}},{"name":"categoria","in":"query","schema":{"$ref":"#/components/schemas/Categoria"}},{"name":"orden","in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Lista de productos","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Producto"}}}}}}}},"/api/productos/{sku}":{"get":{"tags":["Productos"],"summary":"Obtener producto por SKU","parameters":[{"name":"sku","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Producto encontrado","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Producto"}}}},"404":{"description":"No encontrado"}}}},"/api/checkout":{"post":{"tags":["Checkout"],"summary":"Crear orden de compra","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CheckoutRequest"}}}},"responses":{"201":{"description":"Orden creada","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Orden"}}}}}}}},"components":{"schemas":{"LoginRequest":{"type":"object","properties":{"correo":{"type":"string","format":"email"},"password":{"type":"string"}},"required":["correo","password"]},"Usuario":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"rol":{"$ref":"#/components/schemas/Roles"},"createdAt":{"type":"string","format":"date-time"},"correo":{"type":"string","format":"email"},"password":{"type":"string","writeOnly":true},"nombre":{"type":"string"},"rut":{"type":"string"},"telefono":{"type":"string"},"direccion":{"type":"string"},"pais":{"type":"string"},"region":{"type":"string"},"comuna":{"type":"string"}},"required":["correo","password"]},"Producto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"sku":{"type":"string"},"nombre":{"type":"string"},"precio":{"type":"integer","format":"int64"},"categoria":{"$ref":"#/components/schemas/Categoria"},"fotoUrl":{"type":"string"},"descripcion":{"type":"string"}},"required":["sku","nombre","precio","categoria"]},"CarritoItem":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"sku":{"type":"string"},"cantidad":{"type":"integer","format":"int64"}},"required":["sku","cantidad"]},"ItemOrden":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"productoSku":{"type":"string"},"nombreProducto":{"type":"string"},"precioUnitario":{"type":"integer","format":"int64"},"cantidad":{"type":"integer","format":"int64"}},"required":["productoSku","nombreProducto","precioUnitario","cantidad"]},"Orden":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"numeroOrden":{"type":"string"},"emailCliente":{"type":"string","format":"email"},"usuarioId":{"type":"integer","format":"int64"},"modo":{"type":"string","description":"user o guest"},"fechaCreacion":{"type":"string","format":"date-time"},"total":{"type":"integer","format":"int64"},"estado":{"$ref":"#/components/schemas/EstadoOrden"},"pagoId":{"type":"string"},"codigoAutorizacion":{"type":"string"},"items":{"type":"array","items":{"$ref":"#/components/schemas/ItemOrden"}}},"required":["numeroOrden","emailCliente","total","estado"]},"Categoria":{"type":"string","enum":["MOUSEPAD","MOUSE","AUDIFONOS","TECLADO"]},"Roles":{"type":"string","enum":["CLIENTE","ADMIN"]},"EstadoOrden":{"type":"string","enum":["PAGADO","PREPARANDO","DESPACHADO","ENTREGADO","ANULADO","REEMBOLSADO"]},"CheckoutRequest":{"type":"object","properties":{"usuarioId":{"type":"integer","format":"int64"},"emailInvitado":{"type":"string","format":"email"},"items":{"type":"array","items":{"$ref":"#/components/schemas/CarritoItem"}}},"required":["items"]}}}}